alembic/versions/0002_add_procedure_type.py
"""add procedure_type to appointments

Revision ID: 0002_add_procedure_type
Revises: 0001_initial
Create Date: 2026-02-10 00:00:00.000000
"""

from alembic import op
import sqlalchemy as sa

revision = "0002_add_procedure_type"
down_revision = "0001_initial"
branch_labels = None
depends_on = None


def upgrade() -> None:
    op.add_column("appointments", sa.Column("procedure_type", sa.String(length=50), nullable=True))


def downgrade() -> None:
    op.drop_column("appointments", "procedure_type")
