requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
sqlmodel==0.0.22
alembic==1.13.2
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.9.0
python-dotenv==1.0.1
